backend/requirements.txt
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
pillow==11.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.11.0
python-decouple==3.8
requests==2.32.3
simplejwt==2.0.1
sqlparse==0.5.5
typing==3.7.4.3
typing_extensions==4.15.0
tzdata==2025.3
gunicorn==24.0.0
redis>=5.0.0
celery[redis]>=5.3.0
flower>=2.0.0
google-auth==2.36.0
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.0